+++ CHANGELOG.md
@@ -4,7 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.19] 2019-02-20
+## [0.20] 2019-09-13
+
+### Added
+-  Add `corresponds`/`doesNotCorrespond` to compare values with a correspondence function. This is
+useful when the value doesn't implement `equals`.
+
+### Changed
+- Changed the signature of `isEqualTo` from `Any?` to `T`. This should not effect any existing code
+due to `T` being covariant. The one improvement this brings is that literal numbers can be inferred 
+to the correct type. Before this change `assertThat(1L).isEqualTo(1)` would fail, now it passes.
+
+### Fixed
+- Fixed `isSuccess` failing on null return values
+
+## [0.19] 2019-07-20
 
 ### Fixed
 - Fixed thread-safety issue with soft assertions